lua_Integer to 64-bit, rawgeti() to use lua_Integer
With "stock" settings, lua_Integer is 64-bit even in 32-bit machines. rawgeti() argument n is lua_Integer.

+++ source/derelict/lua/types.d
@@ -92,7 +92,7 @@ enum {
 }
 
 alias lua_Number = double;
-alias lua_Integer = ptrdiff_t;
+alias lua_Integer = long;
 alias lua_Unsigned = uint;
 alias lua_KContext = ptrdiff_t;
 alias LUA_NUMBER = lua_Number;
